+++ tests/unit/test_repositories.py
@@ -1,0 +1,31 @@
+import pytest
+from pip._vendor import html5lib
+
+from pip._internal.repositories.entries import parse_base_url
+
+
+@pytest.mark.parametrize(
+    ("html", "url", "expected"),
+    [
+        (b"<html></html>", "https://example.com/", "https://example.com/"),
+        (
+            b"<html><head>"
+            b"<base href=\"https://foo.example.com/\">"
+            b"</head></html>",
+            "https://example.com/",
+            "https://foo.example.com/",
+        ),
+        (
+            b"<html><head>"
+            b"<base><base href=\"https://foo.example.com/\">"
+            b"</head></html>",
+            "https://example.com/",
+            "https://foo.example.com/",
+        ),
+    ],
+)
+def test_base_url(html, url, expected):
+    document = html5lib.parse(
+        html, transport_encoding=None, namespaceHTMLElements=False,
+    )
+    assert parse_base_url(document, url) == expected
